Split off importform
